+++ KarhooSDKTests/TestCases/Service/Loyalty/KarhooLoyaltyServiceSpec.swift
@@ -18,7 +18,7 @@ final class KarhooLoyaltyServiceSpec: XCTestCase {
     private var mockLoyaltyBurnInteractor: MockLoyaltyBurnInteractor!
     private var mockLoyaltyEarnInteractor: MockLoyaltyEarnInteractor!
     private var mockLoyaltyPreAuthInteractor: MockLoyaltyPreAuthInteractor!
-    private var mockLoyaltyRefreshCurrentStatusInteractor: MockLoyaltyRefreshCurrentStatusInteractor!
+    private var mockUserDataStore: MockUserDataStore!
     
     private let identifier = "some_id"
     private let currency = "GBP"
@@ -35,20 +35,20 @@ final class KarhooLoyaltyServiceSpec: XCTestCase {
     override func setUp() {
         super.setUp()
         
+        mockUserDataStore = MockUserDataStore()
         mockLoyaltyBalanceInteractor = MockLoyaltyBalanceInteractor()
         mockLoyaltyConversionInteractor = MockLoyaltyConversionInteractor()
         mockLoyaltyStatusInteractor = MockLoyaltyStatusInteractor()
         mockLoyaltyBurnInteractor = MockLoyaltyBurnInteractor()
         mockLoyaltyEarnInteractor = MockLoyaltyEarnInteractor()
         mockLoyaltyPreAuthInteractor = MockLoyaltyPreAuthInteractor()
-        mockLoyaltyRefreshCurrentStatusInteractor = MockLoyaltyRefreshCurrentStatusInteractor()
-        testObject = KarhooLoyaltyService(loyaltyBalanceInteractor: mockLoyaltyBalanceInteractor,
+        testObject = KarhooLoyaltyService(userDataStore: mockUserDataStore,
+                                          loyaltyBalanceInteractor: mockLoyaltyBalanceInteractor,
                                           loyaltyConversionInteractor: mockLoyaltyConversionInteractor,
                                           loyaltyStatusInteractor: mockLoyaltyStatusInteractor,
                                           loyaltyBurnInteractor: mockLoyaltyBurnInteractor,
                                           loyaltyEarnInteractor: mockLoyaltyEarnInteractor,
-                                          loyaltyPreAuthInteractor: mockLoyaltyPreAuthInteractor,
-                                          loyaltyRefreshCurrentStatusInteractor: mockLoyaltyRefreshCurrentStatusInteractor)
+                                          loyaltyPreAuthInteractor: mockLoyaltyPreAuthInteractor)
     }
     
     /**
@@ -207,31 +207,6 @@ final class KarhooLoyaltyServiceSpec: XCTestCase {
         let expectedError = TestUtil.getRandomError()
         mockLoyaltyPreAuthInteractor.triggerFail(error: expectedError)
 
-        XCTAssert(expectedError.equals(result?.errorValue()))
-    }
-    
-    func testLoyaltyRefreshStatusSuccess() {
-        let call = testObject.refreshCurrentLoyaltyStatus(identifier: identifier)
-        
-        var result: Result<LoyaltyStatus>?
-        call.execute(callback: { result = $0 })
-        
-        mockLoyaltyRefreshCurrentStatusInteractor.triggerSuccess(result: loyaltyStatusMock)
-        
-        XCTAssertEqual(1000, result?.successValue()?.balance)
-        XCTAssertEqual(false, result?.successValue()?.canBurn)
-        XCTAssertEqual(true, result?.successValue()?.canEarn)
-    }
-    
-    func testLoyaltyRefreshStatusFail() {
-        let call = testObject.refreshCurrentLoyaltyStatus(identifier: identifier)
-        
-        var result: Result<LoyaltyStatus>?
-        call.execute(callback: { result = $0 })
-        
-        let expectedError = TestUtil.getRandomError()
-        mockLoyaltyRefreshCurrentStatusInteractor.triggerFail(error: expectedError)
-        
         XCTAssert(expectedError.equals(result?.errorValue()))
     }
 }
